+++ src/index.js
@@ -1,16 +1,18 @@
 import data from "./data";
 import * as visualizers from "./visualizers";
 import properties from "./properties";
+import utils from "./utils";
 
 // External libraries
 import '../node_modules/bootstrap/dist/css/bootstrap.min.css';
 import '../node_modules/font-awesome/css/font-awesome.min.css';
 
-const version = '0.2.5';
+const version = '0.2.6';
 
 export {
     version,
     data,
     visualizers,
     properties,
+    utils,
 };