+++ src/utils/index.js
@@ -1,0 +1,7 @@
+const UrlUtils = require("./Url.utils");
+
+const utils = {
+    url: UrlUtils,
+}
+
+module.exports = utils;